+++ official/utils/flags/core.py
@@ -81,7 +81,6 @@ define_performance = register_key_flags_in_core(_performance.define_performance)
 help_wrap = _conventions.help_wrap
 
 
-apply_clean = _base.apply_clean
 get_num_gpus = _base.get_num_gpus
 get_tf_dtype = _performance.get_tf_dtype
 get_loss_scale = _performance.get_loss_scale
